PIM-10607: Only request /announcements when the panel is open
