+++ src/Akeneo/Platform/Bundle/CommunicationChannelBundle/front/src/components/panel/AnnouncementList.tsx
@@ -25,10 +25,6 @@ const AnnouncementList = ({campaign, panelIsClosed}: ListAnnouncementProps) => {
   const handleHasNewAnnouncements = useHasNewAnnouncements();
   const handleAddViewedAnnouncements = useAddViewedAnnouncements();
 
-  useEffect(() => {
-    handleHasNewAnnouncements();
-  }, []);
-
   const updateNewAnnouncements = useCallback(async () => {
     const newAnnouncements = announcementResponse.items.filter((item: Announcement) => item.tags.includes('new'));
     if (newAnnouncements.length > 0) {
@@ -50,6 +46,10 @@ const AnnouncementList = ({campaign, panelIsClosed}: ListAnnouncementProps) => {
 
   if (announcementResponse.hasError) {
     return <EmptyAnnouncementList text={__('akeneo_communication_channel.panel.list.error')} />;
+  }
+
+  if (announcementResponse.isFetching && announcementResponse.items.length === 0) {
+    return null;
   }
 
   if (announcementResponse.items.length === 0) {
